tests/wpt/tests/html/interaction/focus/focus-file-input.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: file input can be programatically focused before layout</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#focus">
<link rel="help" href="https://bugzilla.mozilla.org/show_bug.cgi?id=505355">
<link rel="author" title="Mozilla" href="https://mozilla.org/">
<link rel="author" title="Emilio Cobos Ãlvarez" href="mailto:emilio@crisal.io">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<input type="file">
<script>
test(function() {
  let input = document.querySelector("input");
  assert_not_equals(document.activeElement, input);
  input.focus();
  assert_equals(document.activeElement, input);
});
</script>
